split/fix IGField.set_dofs(), new Field.set_dofs()
- split IGField.set_dofs() into new IGField.get_surface_basis()
  (IGField-specific) and project_to_facets() (general)
- fixes wrong shape of projected function return value
- Field.set_dofs() calls project_to_facets()
